Split the net system state and the two largest command files
  NetSyncSystem.cs (741) -> the queues and geometry facts; ApplyState.cs, the
    state the commit machinery keeps, next to the Apply*.cs files that read it;
    Lifecycle.cs
  ObjectToolOperationCommand.cs (730) -> the ten portable intent types move to
    ObjectToolIntents.cs, leaving the command that carries them
  ResidentialOccupancySnapshot.cs (726) -> the five record structs to
    OccupancyRecords.cs, page validation and the name table to
    OccupancySnapshotValidation.cs

Also refreshes NetSyncSystem's class comment, which still described a file
layout from before this pass.

Verified by comparing every type body and every class member against the
previous revision, plus a repo-wide check that no type is now declared twice.
